pyproject.toml
[project]
name = "AI-Agent"
version = "0.1.0"
description = "A simple project to build an AI-Agent using Gemini free API."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

